use xz compression for fastboot
When building a images for memory-restricted systems (e.g. 128MB RAM or less), it's important to save space. Using xz compression instead of gzip significantly reduces the size of the fastboot squashfs, reducing the time needed to transfer it from the boot medium or TFTP server and saving lots of memory.