Update Pkgfile
https://github.com/rdesktop/rdesktop/releases/tag/v1.8.5

This is a security release to address various buffer overflow and overrun issues in the rdesktop protocol handling. rdesktop will now detect any attempts to access invalid areas and refuse to continue. Users are adviced to upgrade as soon as possible.

A big thank you to Kaspersky Lab and National Cyber Security Centre for identifying these issues.